pyproject.toml
[tool.poetry]
name = "budgetting"
version = "0.1.0"
description = ""
authors = ["Dylan Kirby <dylankirbydev@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
